fixes entry for PR #904 - duplicate invlet
Pull request from entrez:  explicitly throwing 1 out of stack of more
than 1 and then having the throw be rejected by tageting yourself did
not recombine the split, resulting in stacks of 1 and N-1 that both
had the same inventory letter.  Undo the split if throwing fails.

A similar fix was made for gold a year and a half ago by commit
2d2b2ba95f5945825007c3acf3d0d44187cb9f45.

fixes #904